Polynomials

TensorDec.hilbert

TensorDec.perp

Missing docstring.

Missing docstring for @ring. Check Documenter's build log for details.

TensorDec.perp — Function

Compute the kernel of the Hankel matrix in degree d of the symmetric tensor F.

TensorDec.hilbert — Function

Sequence of dimension of $S/(F^⟂)$ or of the kernels of the Hankel matrix in degree i for i in 1:deg(F).