Series

MultivariateSeries.dual

MultivariateSeries.hankel

MultivariateSeries.dual — Function

dual(p::Polynomial, d:: Int64) -> Series{T}

Compute the series associated to the tensor p of degree d. T is the type of the coefficients of the polynomial p.

MultivariateSeries.hankel — Function

Compute the Hankel matrix (a.k.a. Catalecticant matrix) in degree d of the symmetric tensor F.