Polynomials

TensorDec.ahp

TensorDec.hilbert

TensorDec.hpol

TensorDec.perp

TensorDec.hpol — Function

hpol(W,A,X,d) ⤍ Homogeneous polynomial

This function gives the homogeneous polynomial associated to the symmetric decomposition W,A.

```

TensorDec.ahp — Function

ahp(T::symmetric Tensor, X=@polyvar x1...xn)-> 'P' Associated homogeneous polynomial

The associated homogeneous polynomials of degree d in n variables of a symmetric tensor of order d and dimension n.

Example

julia> n=2
2

julia> d=3
3

julia> T
2×2×2 Array{Float64,3}:
[:, :, 1] =
 -3.0  -1.5
 -1.5   0.0

[:, :, 2] =
 -1.5  0.0
  0.0  1.5

julia> X=@polyvar x1 x2
2-element Array{PolyVar{true},1}:
 x1
 x2

 julia> P=ahp(T,X)
 (-3.0 + 0.0im)x1³ + (-4.5 + 0.0im)x1²x2 + (1.5 + 0.0im)x2³

TensorDec.perp — Function

Compute the kernel of the Hankel matrix in degree d of the symmetric tensor F.

TensorDec.hilbert — Function

Sequence of dimension of $S/(F^⟂)$ or of the kernels of the Hankel matrix in degree i for i in 1:maxdegree(F).